#50 Refactor codes  This commit will refactor the endpoints into a more reasonable controllers.  Also, new postman tests have been created.
