+++ src/app.js
@@ -7,6 +7,7 @@ const UniversityController = require("./controllers/university-controller");
 const ProgramController = require("./controllers/programs-controller");
 const UserController = require("./controllers/user-controller");
 const ConsentController = require("./controllers/consentForm-controller");
+const AuthController = require("./controllers/auth-controller");
 
 const app = express();
 require('dotenv').config()
@@ -48,6 +49,7 @@ app.use("/api/universities", UniversityController)
 app.use("/api/programs", ProgramController);
 app.use("/api/users", UserController);
 app.use("/api/consents", ConsentController);
+app.use("/api/auth", AuthController)
 
 app.listen(port, function (err) {
   if (err) throw err;
